+++ crates/types-traits/domain_types/src/payment_address.rs
@@ -435,6 +435,12 @@ pub struct OrderDetailsWithAmount {
     pub total_amount: Option<MinorUnit>,
     /// discount amount on the unit
     pub unit_discount_amount: Option<MinorUnit>,
+    // discount name applied to this item
+    pub discount_name: Option<String>,
+    // discount percentage applied to this item
+    pub discount_percentage: Option<f64>,
+    // Discount type applied to this item
+    pub discount_type: Option<String>,
 }
 
 impl SerializableSecret for OrderDetailsWithAmount {}